% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres_grch37}
\alias{centromeres_grch37}
\title{Centromeres grch37}
\format{
## `centromeres_grch37`
A data frame with 22 rows and 4 columns.
\describe{
  \item{chrom}{Factor. Chromosomes without chr-prefix, 1:22.}
  \item{start}{Integer. Start coordinates for the specified chromosome arm.}
  \item{end}{Integer. End coordinates for the specified chromosome arm.}
  \item{arm}{Factor. Chromosome arm, (always centromere).}
}
}
\usage{
centromeres_grch37
}
\description{
A data frame with the centromeres coordinates in respect to grch37.
}
\keyword{datasets}
