% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres_hg38}
\alias{centromeres_hg38}
\title{Centromeres hg38}
\format{
## `centromeres_hg38`
A data frame with 22 rows and 4 columns.
\describe{
  \item{chrom}{Factor. chr-prefixed chromosomes, chr1-chr22.}
  \item{start}{Integer. Start coordinates for the specified chromosome arm.}
  \item{end}{Integer. End coordinates for the specified chromosome arm.}
  \item{arm}{Factor. Chromosome arm (always centromere).}
}
}
\usage{
centromeres_hg38
}
\description{
A data frame with the centromeres coordinates in respect to hg38.
}
\keyword{datasets}
