% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome_arms_hg38}
\alias{chromosome_arms_hg38}
\title{Chromosome Arms hg38.}
\format{
## `chromosome_arms_hg38`
A data frame with 44 rows and 4 columns.
\describe{
  \item{chrom}{Factor. chr-prefixed chromosomes, chr1-chr22.}
  \item{start}{Integer. Start coordinates for the specified chromosome arm.}
  \item{end}{Integer. End coordinates for the specified chromosome arm.}
  \item{arm}{Factor. Chromosome arm, either p or q.}
}
}
\usage{
chromosome_arms_hg38
}
\description{
A data frame with the chromosome arm coordinates in respect to hg38.
}
\keyword{datasets}
