% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_ranger.R
\name{cyto_ranger}
\alias{cyto_ranger}
\title{Cyto Ranger.}
\usage{
cyto_ranger(
  these_regions = NULL,
  qchrom = NULL,
  qstart = NULL,
  qend = NULL,
  projection = "hg38"
)
}
\arguments{
\item{these_regions}{The region(s) to be queried. Can be a data frame with
regions with the following columns; chrom, start, end.
Or in a string in the following format chr:start-end.}

\item{qchrom}{Query chromosome (prefixed or un-prefixed),
Required if `these_regions` is not provided.}

\item{qstart}{Query start position. Required if `these_regions` is not provided.}

\item{qend}{Query end position. Required if `these_regions` is not provided.}

\item{projection}{The desired projection you want back coordinates for.
Available projections are hg38 and grch37. Default is hg38.}
}
\value{
A data frame with cytoband information for the specified region(s).
}
\description{
This function takes a region or regions and returns the cytoband
information for the region(s).
}
\details{
Query a region and relevant cytoband information for the specified 
region(s). This function accepts a variety of incoming regions.
Either, regions can be provided as a data frame with `these_regions`.
If so, the following columns must exist; chrom, start, end.
This parameter (`these_regions`) also accept a region in "region" format,
(i.e chr:start-end). This can be a region or a vector of characters with
multiple regions. The user can also individually specify region(s) with;
`qchrom` (string), `qstart` (string, or integer), and `qend` (string or integer).
These parameters can also accept a vector of characters for multiple regions.
The function also handles chromosome prefixes in the returned object,
based on the selected `projection`.
}
\examples{
#' #Example 1 - Give the function one region as a string
cyto_ranger(these_regions = "chr8:127735434-127742951")

#Example 2 - Give the function multiple regions as a string
cyto_ranger(these_regions = c("chr8:128747680-128753674",
                              "chr18:60790579-60987361"),
            projection = "grch37")

#Example 3 - Individually specify the chromosome, start and end coordinates
cyto_ranger(qchrom = "chr8",
            qstart = 127735434,
            qend = 127742951)

#Example 4 - Individually specify multiple regions with the query parameters
cyto_ranger(qchrom = c("chr8", "chr18"),
            qstart = c(128747680, 60790579),
            qend = c(128753674, 60987361),
            projection = "grch37")

}
