% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands_grch37}
\alias{cytobands_grch37}
\title{Cytobands grch37}
\format{
## `cytobands_grch37`
A data frame with 862 rows and 5 columns.
\describe{
  \item{chrom}{Factor. Chromosomes without chr-prefix, 1:22.}
  \item{start}{Integer. Start coordinates for the specified chromosome arm.}
  \item{end}{Integer. End coordinates for the specified chromosome arm.}
  \item{cytoBand}{Character. Cytoband name for the specified range.}
  \item{gieStain}{Factor. Giemsa Stain proporties for the described cyto band.}
}
}
\usage{
cytobands_grch37
}
\description{
A data frame with the cytobands coordinates in respect to grch37.
}
\keyword{datasets}
