% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands_hg38}
\alias{cytobands_hg38}
\title{Cytobands hg38}
\format{
## `cytobands_hg38`
A data frame with 1549 rows and 5 columns.
\describe{
  \item{chrom}{Factor. chr-prefixed chromosomes, chr1-chr22.}
  \item{start}{Integer. Start coordinates for the specified chromosome arm.}
  \item{end}{Integer. End coordinates for the specified chromosome arm.}
  \item{cytoBand}{Character. Cytoband name for the specified range.}
  \item{gieStain}{Factor. Giemsa Stain proporties for the described cyto band.}
}
}
\usage{
cytobands_hg38
}
\description{
A data frame with the cytobands coordinates in respect to hg38.
}
\keyword{datasets}
