% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_coordinates_grch37}
\alias{gene_coordinates_grch37}
\title{Gene Coordinates grch37.}
\format{
## `gene_coordinates_grch37`
A data frame with 57905 rows and 5 columns.
\describe{
  \item{chrom}{Factor. The chromosome (un-prefixed) that the gene is residing on.}
  \item{start}{Integer. The start coordinates for the gene.}
  \item{end}{Integer. The end coordinates for the gene.}
  \item{width}{Integer. The width for the gene.}.
  \item{strand}{Defined as + forward or - reverse.}.
  \item{type}{Feature type name. Current allowed features are; gene, transcript, exon, CDS, Selenocysteine, start_codon, stop_codon and UTR}.
  \item{tag}{A collection of additional key value tags}.
  \item{ccds_id}{CCDS identifier linked to this transcript}.
  \item{ensembl_gene_id}{Character. Ensembl gene ID.}
  \item{hugo_symbol}{Character. Gene symbol in Hugo format.}
  \item{source}{Name of the program that generated this feature, or the data source (database or project name)}.
  \item{score}{A floating point value indiciating the score of a feature}.
  \item{gene_version}{The stable identifier version for the gene}.
  \item{gene_source}{The annotation source for this gene}.
  \item{gene_biotype}{The biotype of this gene}.
  \item{transcript_id}{The stable identifier for this transcript}.
  \item{transcript_version}{The stable identifier version for this transcript}.
  \item{transcript_name}{The symbold for this transcript derived from the gene name}.
  \item{transcript_source}{The annotation source for this transcript}.
  \item{transcript_biotype}{The biotype for this transcript}.
  \item{exon_number}{Position of this exon in the transcript}.
  \item{exon_id}{The stable identifier for this exon}.
  \item{exon_version}{The stable identifier version for this exon}.
  \item{protein_id}{Stable identifier for this transcript's protein}.
  \item{protein_version}{Stable identifier version for this transcript's protein}.
}
}
\usage{
gene_coordinates_grch37
}
\description{
All gene coordinates in respect to grch37. This data set is downloaded from ensembl.
}
\keyword{datasets}
