% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Get Gene Info}
\usage{
get_gene_info(these_genes = NULL, projection = "hg38", raw = FALSE)
}
\arguments{
\item{these_genes}{Required argument. The gene or genes of interest.}

\item{projection}{The desired projection, default is hg38.}

\item{raw}{Default is FALSE, set to TRUE for keeping all columns.}
}
\value{
A data frame with gene information.
}
\description{
Convenience function for returning information about a gene or
a set of genes. This function is internally called by
[BioMaesteR::get_gene_region].
}
\details{
Give the function a gene or a set of genes (as a vector of characters),
specify the projection (if not, hg38 is the default projection) and return
gene information based on the bundled data. By default this function is
run with `raw = FALSE`, this returns a subset of columns. If instead the user
wants everything back (i.e all available columns) toggle `raw` to `TRUE`.
}
\examples{
#Example 1 - Query one gene (in Hugo format) and with default parameters.
get_gene_info(these_genes = "MYC")

#Example 2 - Same as example 1 but MYC is here specified as Ensembl ID.
get_gene_info(these_genes = "ENSG00000136997")

#Example 3 - Request multiple genes with non-default parameters
get_gene_info(these_genes = c("MYC", "BCL2"),
              projection = "grch37")

#Example 4 - Request multiple Ensembl IDs and return all columns.
get_gene_info(these_genes = c("ENSG00000136997", "ENSG00000171791"),
              raw = TRUE)

}
