% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify_chr.R
\name{purify_chr}
\alias{purify_chr}
\title{Purify Chromosomes.}
\usage{
purify_chr(projection = NULL, incoming_table = NULL)
}
\arguments{
\item{projection}{Required parameter, needed to determine if chromosomes 
should be prefixed or not.}

\item{incoming_table}{Required parameter, a data frame to check for prefixes.}
}
\value{
A data frame with the same columns as the incoming data frame, but
with the prefixes added or removed based on the selected projection.
}
\description{
Helper function for dealing with chromosome prefixes based on 
the selected projection.
}
\details{
This function accepts a data frame with a column to check for
prefixes. The function will add or remove prefixes based on the selected
projection. This function expects a column named "chrom" to be in the 
incoming data frame.
}
\examples{
#Example 1 - Add prefixes to a data frame
my_data = data.frame(chrom = c("1", "2", "3"))
my_data = purify_chr(projection = "hg38", incoming_table = my_data)

}
